# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLMNightRun",
    "description": "멀티 LLM 통합 자동화 플랫폼",
    "version": "0.1.0",
    "x-logo": {
      "url": "/static/logo.png",
      "altText": "LLMNightRun"
    }
  },
  "paths": {
...
    "/questions/": {
      "post": {
        "tags": [
          "questions"
        ],
        "summary": "Submit Question",
        "description": "새로운 질문을 데이터베이스에 저장합니다.\n\nArgs:\n    question: 질문 내용 및 태그\n    db: 데이터베이스 세션\n    \nReturns:\n    QuestionResponse: 생성된 질문 정보\n    \nRaises:\n    HTTPException: 질문 생성 중 오류 발생 시",
        "operationId": "submit_question_questions__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "404": {
            "description": "질문을 찾을 수 없음"
          },
          "500": {
            "description": "서버 내부 오류"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "questions"
        ],
        "summary": "List Questions",
        "description": "질문 목록을 페이지네이션과 필터링 옵션으로 조회합니다.\n\nArgs:\n    skip: 건너뛸 항목 수\n    limit: 최대 조회 항목 수\n    tag: 필터링할 태그 (선택 사항)\n    db: 데이터베이스 세션\n    \nReturns:\n    List[QuestionResponse]: 질문 목록\n    \nRaises:\n    HTTPException: 질문 목록 조회 중 오류 발생 시",
        "operationId": "list_questions_questions__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponse"
                  },
                  "title": "Response List Questions Questions  Get"
                }
              }
            }
          },
          "404": {
            "description": "질문을 찾을 수 없음"
          },
          "500": {
            "description": "서버 내부 오류"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          }
        }
      }
    },
    "/questions/{question_id}": {
      "get": {
        "tags": [
          "questions"
        ],
        "summary": "Get Question",
        "description": "특정 ID의 질문을 조회합니다.\n\nArgs:\n    question_id: 질문 ID\n    db: 데이터베이스 세션\n    \nReturns:\n    QuestionResponse: 질문 정보\n    \nRaises:\n    HTTPException: 질문을 찾을 수 없거나 조회 중 오류 발생 시",
        "operationId": "get_question_questions__question_id__get",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "404": {
            "description": "질문을 찾을 수 없음"
          },
          "500": {
            "description": "서버 내부 오류"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "questions"
        ],
        "summary": "Update Question Endpoint",
        "description": "특정 ID의 질문을 업데이트합니다.\n\nArgs:\n    question_id: 질문 ID\n    question: 업데이트할 질문 정보\n    db: 데이터베이스 세션\n    \nReturns:\n    QuestionResponse: 업데이트된 질문 정보\n    \nRaises:\n    HTTPException: 질문을 찾을 수 없거나 업데이트 중 오류 발생 시",
        "operationId": "update_question_endpoint_questions__question_id__put",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Question"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "404": {
            "description": "질문을 찾을 수 없음"
          },
          "500": {
            "description": "서버 내부 오류"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "questions"
        ],
        "summary": "Delete Question Endpoint",
        "description": "특정 ID의 질문을 삭제합니다.\n\nArgs:\n    question_id: 질문 ID\n    db: 데이터베이스 세션\n    \nReturns:\n    Dict[str, Any]: 삭제 성공 메시지\n    \nRaises:\n    HTTPException: 질문을 찾을 수 없거나 삭제 중 오류 발생 시",
        "operationId": "delete_question_endpoint_questions__question_id__delete",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Question Endpoint Questions  Question Id  Delete"
                }
              }
            }
          },
          "404": {
            "description": "질문을 찾을 수 없음"
          },
          "500": {
            "description": "서버 내부 오류"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          }
        }
      }
    },
    "/questions/search/": {
      "get": {
        "tags": [
          "questions"
        ],
        "summary": "Search Questions Endpoint",
        "description": "질문을 검색어를 기준으로 검색합니다.\n\nArgs:\n    query: 검색어 (최소 2글자)\n    skip: 건너뛸 항목 수\n    limit: 최대 조회 항목 수\n    db: 데이터베이스 세션\n    \nReturns:\n    List[QuestionResponse]: 검색된 질문 목록\n    \nRaises:\n    HTTPException: 검색 중 오류 발생 시",
        "operationId": "search_questions_endpoint_questions_search__get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "description": "검색어",
              "title": "Query"
            },
            "description": "검색어"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponse"
                  },
                  "title": "Response Search Questions Endpoint Questions Search  Get"
                }
              }
            }
          },
          "404": {
            "description": "질문을 찾을 수 없음"
          },
          "500": {
            "description": "서버 내부 오류"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          }
        }
      }
    },
...
    "/export/question/{question_id}": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Question",
        "description": "질문과 관련된 모든 응답 및 코드 스니펫을 내보냅니다.\n\nArgs:\n    question_id: 질문 ID\n    format: 내보내기 형식 (markdown, json, html, pdf, code_package)\n    include_metadata: 메타데이터 포함 여부\n    include_tags: 태그 포함 여부\n    include_timestamps: 타임스탬프 포함 여부\n    include_llm_info: LLM 정보 포함 여부\n    code_highlighting: 코드 하이라이팅 여부\n    db: 데이터베이스 세션\n    \nReturns:\n    형식에 맞는 내보내기 파일",
        "operationId": "export_question_export_question__question_id__get",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "내보내기 형식",
              "default": "markdown",
              "title": "Format"
            },
            "description": "내보내기 형식"
          },
          {
            "name": "include_metadata",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "메타데이터 포함 여부",
              "default": true,
              "title": "Include Metadata"
            },
            "description": "메타데이터 포함 여부"
          },
          {
            "name": "include_tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "태그 포함 여부",
              "default": true,
              "title": "Include Tags"
            },
            "description": "태그 포함 여부"
          },
          {
            "name": "include_timestamps",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "타임스탬프 포함 여부",
              "default": true,
              "title": "Include Timestamps"
            },
            "description": "타임스탬프 포함 여부"
          },
          {
            "name": "include_llm_info",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "LLM 정보 포함 여부",
              "default": true,
              "title": "Include Llm Info"
            },
            "description": "LLM 정보 포함 여부"
          },
          {
            "name": "code_highlighting",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "코드 하이라이팅 여부",
              "default": true,
              "title": "Code Highlighting"
            },
            "description": "코드 하이라이팅 여부"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "리소스를 찾을 수 없음"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Question": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "Question",
        "description": "사용자 질문 모델"
      },
      "QuestionCreate": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "QuestionCreate",
        "description": "질문 생성 요청 모델"
      },
      "QuestionResponse": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "QuestionResponse",
        "description": "질문 응답 모델 (API 응답용)"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
  "servers": [
    {
      "url": "",
      "description": "개발 서버"
    }
  ]
}